asset bug fix
